requirements.txt
Appium-Python-Client==0.22
selenium==2.53.2
 